extract utils & fix highlight width
