+++ package.json
@@ -39,6 +39,7 @@
     "d3-interpolate-path": "^2.2.3",
     "d3-scale": "^2",
     "d3-shape": "^3.0.1",
+    "react-keyed-flatten-children": "^1.3.0",
     "react-native-redash": "^16.1.1",
     "react-native-svg": "^12.1.1"
   },
