+++ src/charts/line/Group.tsx
@@ -1,8 +1,9 @@
 import React from 'react';
 import { ReactNode, Children, cloneElement } from 'react';
 import { ViewProps, View } from 'react-native';
+import flattenChildren from 'react-keyed-flatten-children';
+
 import { LineChart } from './Chart';
-import { flattenChildren } from './utils';
 
 type Props = {
   children: ReactNode;
