+++ src/charts/line/utils/index.ts
@@ -1,0 +1,5 @@
+export * from './getArea';
+export * from './getDomain';
+export * from './getPath';
+export * from './interpolatePath';
+export * from './lineChartDataPropToArray';
