+++ src/index.ts
@@ -1,4 +1,4 @@
 export * from './charts/candle';
 export * from './charts/line';
-export * from './utils';
 export * from './components/AnimatedText';
+export * from './utils';
