+++ src/utils/formatPrice.ts
@@ -1,5 +1,3 @@
-import * as React from 'react';
-
 /**
  * @worklet
  */
@@ -39,33 +37,4 @@ export function formatPrice({
     }
   }
   return res;
-}
-
-/**
- * @worklet
- */
-export function formatDatetime({
-  value,
-  locale = 'en-US',
-  options = {},
-}: {
-  value: number;
-  locale?: string;
-  options?: Intl.DateTimeFormatOptions;
-}) {
-  'worklet';
-  const d = new Date(value);
-  return d.toLocaleString(locale, options);
-}
-
-export function usePrevious<T>(value: T) {
-  // The ref object is a generic container whose current property is mutable ...
-  // ... and can hold any value, similar to an instance property on a class
-  const ref = React.useRef<T>();
-  // Store current value in ref
-  React.useEffect(() => {
-    ref.current = value;
-  }, [value]); // Only re-run if value changes
-  // Return previous value (happens before update in useEffect above)
-  return ref.current;
 }
